[00018] Move PluginManagerApp to core framework, registered in UsePlugins
