--- src/plugins/plugins/Ivy.Plugin.PluginManager/Ivy.Plugin.PluginManager.csproj
@@ -1,12 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>net10.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="../../../Ivy.Plugin.Abstractions/Ivy.Plugin.Abstractions.csproj" />
-    <ProjectReference Include="../../../Ivy/Ivy.csproj" />
-  </ItemGroup>
-</Project>
